+++ snuba/datasets/storages/groupedmessages.py
@@ -7,7 +7,7 @@ from snuba.datasets.cdc.groupedmessage_processor import (
 from snuba.datasets.cdc.message_filters import CdcTableNameMessageFilter
 from snuba.datasets.dataset_schemas import StorageSchemas
 from snuba.datasets.schemas.tables import ReplacingMergeTreeSchema
-from snuba.datasets.storage import CdcStorage
+from snuba.datasets.cdc import CdcStorage
 from snuba.datasets.table_storage import KafkaStreamLoader, TableWriter
 from snuba.snapshots.loaders.single_table import SingleTableBulkLoader
